Automatic parameter discovery in es_instance_conn_validator
Currently one had to specify the host (which was predictable) but also
the port on which the Elasticsearch instance will be listening to. The
latter was less predictibable. Hence this commit aims to discover
automatically on which port an Elasticsearch instance will be listening
on.

Removed parameters :

  * server
  * port

Mechanism to find out the port : A mix of ps and grep based on the
instance name